refactor: move the block editor out of markdown into its own crate
