+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "crates/ui",
     "crates/syntax",
     "crates/markdown",
+    "crates/editor",
     "crates/terminal",
     "apps/gallery",
     "apps/web",
@@ -33,6 +34,7 @@ theme = { path = "crates/theme", version = "0.0.2" }
 motion = { path = "crates/motion", version = "0.0.2" }
 ui = { path = "crates/ui", version = "0.0.2" }
 markdown = { path = "crates/markdown", version = "0.0.2" }
+editor = { path = "crates/editor", version = "0.0.2" }
 syntax = { path = "crates/syntax", version = "0.0.2" }
 terminal = { path = "crates/terminal", version = "0.0.2" }
