+++ apps/gallery/Cargo.toml
@@ -16,7 +16,8 @@ theme.workspace = true
 
 [dependencies]
 agent.workspace = true
-markdown = { workspace = true, features = ["editor"] }
+editor.workspace = true
+markdown.workspace = true
 motion.workspace = true
 theme.workspace = true
 ui.workspace = true
