+++ crates/editor/src/lib.rs
@@ -1,0 +1,17 @@
+//! A Notion-style block editor for gpui, over the `markdown` document model.
+//!
+//! ```ignore
+//! editor::init(cx);                       // once, at startup
+//! let editor = cx.new(|cx| editor::Editor::new("# Title", cx));
+//! ```
+//!
+//! `markdown` holds the document, its markdown wire form, and the painting —
+//! all of it testable without a window. What lives here is the half that needs
+//! one: focus, keys, the platform input handler, the mouse, undo, and the menus.
+
+mod editor;
+mod history;
+mod slash;
+
+pub use editor::{Editor, EditorHandle, block_label, init};
+pub use history::{DEFAULT_UNDO_LIMIT, EditKind, History};
