+++ crates/markdown/src/lib.rs
@@ -14,29 +14,20 @@
 //! edit/save cycle cannot drift.
 //!
 //! [`doc`], [`parse`], [`serialize`], [`select`] and [`edit`] are pure — no
-//! gpui, no painting — and [`render`] is the gpui layer over them. The editing
-//! *surface* — a focus handle, key bindings and the platform input handler — is
-//! [`editor`], behind the `editor` feature so a read-only consumer compiles
-//! none of it.
+//! gpui, no painting — and [`render`] is the gpui layer over them, caret and
+//! selection included for a caller that owns them. The editing *surface* is the
+//! `editor` crate.
 
 pub mod doc;
 pub mod edit;
-#[cfg(feature = "editor")]
-pub mod editor;
 pub mod highlight;
-#[cfg(feature = "editor")]
-pub mod history;
 pub mod parse;
 pub mod render;
 pub mod select;
 pub mod serialize;
-#[cfg(feature = "editor")]
-pub mod slash;
 
 pub use doc::{Align, Block, BlockKind, Doc, Mark, MarkSpan, Part, Text};
 pub use edit::{Shortcut, shortcut};
-#[cfg(feature = "editor")]
-pub use editor::Editor;
 pub use highlight::{Highlighter, set_highlighter};
 pub use parse::parse;
 pub use render::{BlockLayouts, markdown, render, render_with_selection};
